+++ src/nodes/solution.rs
@@ -9,7 +9,7 @@ impl Solution {
         text: &str,
         url: Option<String>,
         classes: Option<Vec<String>>,
-        forced_level: Option<u32>,
+        forced_level: Option<usize>,
     ) -> EllipticalNode {
         EllipticalNode::new(id, text, None, true, url, classes, forced_level)
     }
