+++ src/nodes/strategy.rs
@@ -10,7 +10,7 @@ impl Strategy {
         undeveloped: bool,
         url: Option<String>,
         classes: Option<Vec<String>>,
-        forced_level: Option<u32>,
+        forced_level: Option<usize>,
     ) -> BoxNode {
         BoxNode::new(id, text, undeveloped, 15, url, classes, forced_level)
     }
